refactor(webview): move main view planning to shared layer
Move browser-safe main view execute-message and action-planning helpers out of controllers, expose the preferred tool-use fallback from shared constants, and route frontend error formatting through shared utils.\n\nCloses #6520
