+++ packages/extension/src/webview/frontend/MainApp.ts
@@ -7,19 +7,6 @@ import '@awesome.me/webawesome/dist/components/icon/icon.js';
 import '@awesome.me/webawesome/dist/components/details/details.js';
 import '@awesome.me/webawesome/dist/components/divider/divider.js';
 import '@awesome.me/webawesome/dist/components/tooltip/tooltip.js';
-import {
-  buildMainViewExecuteMessage,
-  type MainViewExecuteMessage,
-} from '@controllers/mainView/MainViewExecutionMessageController';
-import {
-  planCompare,
-  planLatexdiff,
-  planLatexdiffVC,
-  planLatexdiffVCPack,
-  planMerge,
-  planPackClean,
-} from '@controllers/mainView/MainViewActionPlanner';
-import { PREFERRED_TOOL_USE_AGENTS } from '@agent/index/agentRegistryConstants';
 import { COMMON_COMMANDS, MAIN_VIEW_COMMANDS } from '@shared/ipc';
 import { SignalWatcher, signal, Signal } from '@shared/signals';
 import { BaseWebviewApp } from '@shared/BaseWebviewApp';
@@ -30,6 +17,14 @@ import {
   mainViewMessages,
   MainViewPersistedStateSchema,
   type MainViewMessage,
+  buildMainViewExecuteMessage,
+  planCompare,
+  planLatexdiff,
+  planLatexdiffVC,
+  planLatexdiffVCPack,
+  planMerge,
+  planPackClean,
+  type MainViewExecuteMessage,
   type MainViewPersistedState,
   type ApiKeyBannerState,
   type AgentConfigBannerState,
@@ -64,6 +59,7 @@ import {
   dispatchMainView,
   type MainViewHandlerRegistry,
 } from '@shared/schemas';
+import { PREFERRED_TOOL_USE_AGENTS } from '@shared/constants/agents';
 // Constants-only module: safe for the webview bundle (no platform imports).
 import {
   registerTeXRAWebAwesomeIcons,
