+++ src/controllers/onboarding/setupLaunch.ts
@@ -19,10 +19,9 @@ import {
 } from '@model/setupModelDefaults';
 import { DEFAULT_AGENT_MODEL } from '@shared/constants/providers';
 import { SETUP_AGENT_NAME } from '@shared/constants/agents';
+import type { MainViewExecuteMessage } from '@shared/schemas/mainView';
 import { isNonEmptyString } from '@utils/core';
 import { getUseOpenRouter } from '@utils/config/providerConfig';
-
-import type { MainViewExecuteMessage } from '@controllers/mainView/MainViewExecutionMessageController';
 
 /** Instruction handed to the setup agent on auto-kickoff (mirrors the extension). */
 export const SETUP_INSTRUCTION =
