+++ src/shared/schemas/mainView/index.ts
@@ -24,5 +24,7 @@ export {
   type MultipleDocumentFileType,
 } from '../fileTypes';
 export * from './state';
+export * from './actionPlanner';
+export * from './executeMessage';
 export * from './outbound';
 export * from './inbound';
